types: Move BAP defines to pybtp/defs
This moves the definitions out of btp/bap.py to reuse the values in
other profiles.
